+++ src/main/java/pascal/taie/language/generics/TypeGSignature.java
@@ -20,15 +20,21 @@
  * License along with Tai-e. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package pascal.taie.util.collection;
+
+package pascal.taie.language.generics;
+
 
 import java.io.Serializable;
 
-public record Pair<T1, T2>(T1 first, T2 second)
-        implements Serializable {
-
-    @Override
-    public String toString() {
-        return "<" + first + ", " + second + ">";
-    }
+/**
+ * In <a href="https://docs.oracle.com/javase/specs/jvms/se20/html/jvms-4.html#jvms-JavaTypeSignature">
+ * JVM Spec. 4.7.9.1 JavaTypeSignature</a>,
+ * A <i>Java type signature</i> represents either a reference type
+ * or a primitive type of the Java programming language.
+ * <br>
+ * In our implementation, for convenience, {@link VoidDescriptor} is
+ * also a {@link TypeGSignature}.
+ */
+public sealed interface TypeGSignature extends Serializable
+        permits ReferenceTypeGSignature, BaseType, VoidDescriptor {
 }